publishing and packing Ex3.1

+++ publish_all_examples.m
@@ -3,7 +3,9 @@
 clear; bdclose all; close all; clc
 % Clean up the PUBLISH folder
 pubdir = [pwd '\publish\'];
-rmdir(pubdir,'s');
+if exist('publish', 'dir')
+    rmdir(pubdir,'s');
+end
 mkdir(pubdir);
 
 %% Publish Example 2.1
@@ -22,5 +24,15 @@ options = struct('outputDir',[cpath '\publish\Ch2Ex2\']);
 cd('Chapter2_Example2')
 publish('Ch2Ex2_main.m', options);
 movefile([cpath '\publish\Ch2Ex2\Ch2Ex2_main.html'], [cpath '\publish\Ch2Ex2\index.html']);
+cd(cpath)
+clear; close all;  % clean up
+
+%% Publish Example 3.1
+disp('-->Publishing Example 3.1')
+cpath = pwd; % Save the root level path
+options = struct('outputDir',[cpath '\publish\Ch3Ex1\']);
+cd('Chapter3_Example1')
+publish('Ch3Ex1_main.m', options);
+movefile([cpath '\publish\Ch3Ex1\Ch3Ex1_main.html'], [cpath '\publish\Ch3Ex1\index.html']);
 cd(cpath)
 clear; close all;  % clean up
